swapix api/requirements.txt
albumentations==1.4.7
annotated-types==0.6.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
easydict==1.13
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
prettytable==3.10.0
protobuf==5.26.1
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.5.10
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
